requirements.txt
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
croniter==6.0.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-rq==3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
rq==2.6.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.10.0
